Refactor : Move hooks to hook folder
