+++ src/views/List/StockItem/SummaryInfo/hooks/useChangeStockCheckbox.tsx
@@ -1,0 +1,23 @@
+import { useDispatch } from 'react-redux';
+import { updateCheckedItems } from '../../../../../features/checkedItems/checkedItemsSlice';
+import { useCallback } from 'react';
+
+const useChangeStockCheckbox = (stockId: string) => {
+  const dispatch = useDispatch();
+
+  const onChangeCheckbox = useCallback(
+    (value: boolean) => {
+      dispatch(
+        updateCheckedItems({
+          type: 'stock',
+          checked: value,
+          stockId: stockId,
+        }),
+      );
+    },
+    [dispatch, stockId],
+  );
+  return onChangeCheckbox;
+};
+
+export default useChangeStockCheckbox;
